+++ examples/acquired_immunity/acquired_immunity.jl
@@ -23,8 +23,8 @@ function run(seed::Int64, t_vec::Vector{Float64})
         "pat_type",
         num_loci=4,
         possible_alleles="AB",
-        mean_effective_inoculum=1.0,
-        mean_mutations_per_replication=0.00003,
+        # mean_effective_inoculum=1.0,
+        # mean_mutations_per_replication=0.00003,
         contactCoefficient=s::String -> 1.0 + (0.1 * (4.0 - hamming(s, optimal_genome)) / 4.0),
         receiveContactCoefficient=s::String -> 0.0,
     )
@@ -32,7 +32,7 @@ function run(seed::Int64, t_vec::Vector{Float64})
     res_type = jOpqua.newResponseType(
         "res_type",
         reactivityCoefficient=(hos_g::String, imp_g::String, mat_g::String, pat_g::String) -> imp_g == pat_g ? 1.0 : 0.0,
-        infectionCoefficient=(hos_g::String, imp_g::String, mat_g::String, pat_g::String) -> imp_g == pat_g ? 0.0 : 1.0,
+        transmissionEfficiencySpecificCoefficient=(hos_g::String, imp_g::String, mat_g::String, pat_g::String) -> imp_g == pat_g ? 0.0 : 1.0,
         clearanceSpecificCoefficient=(hos_g::String, imp_g::String, mat_g::String, pat_g::String) -> imp_g == pat_g ? 1.0e3 : 1.0,
         # responseAcquisitionSpecificCoefficient=(hos_g::String, imp_g::String, mat_g::String, pat_g::String) -> imp_g == pat_g ? 0.0 : 1.0,
     )
@@ -46,6 +46,8 @@ function run(seed::Int64, t_vec::Vector{Float64})
         response_acquisition_coefficient=0.5,
         response_loss_coefficient=0.0,
         receive_contact_coefficient=1.0,
+        mutations_upon_infection_coefficient=0.00003,
+        inoculum_coefficient=1.0,
         pathogenFractions=jOpqua.pathogenFractionsProportionalFitness,
         response_types=Dict{String,jOpqua.ResponseType}([(res_type.id => res_type)]),
         developResponses=(