+++ src/parameters/default.jl
@@ -24,7 +24,7 @@ const DEFAULT_RESPONSE_TYPE = ResponseType(
     0.0,
     # (hos_g::String, imp_g::String, mat_g::String, pat_g::String) -> 1.0,
     (hos_g::String, imp_g::String, mat_g::String, pat_g::String) -> 1.0,
-    (hos_g::String, imp_g::String, mat_g::String, pat_g::String) -> 1.0,
+    # (hos_g::String, imp_g::String, mat_g::String, pat_g::String) -> 1.0,
     (hos_g::String, imp_g::String, mat_g::String, pat_g::String) -> 1.0,
     SA[ # order defined in COEFFICIENTS
         (hos_g::String, imp_g::String, mat_g::String)->1.0, (hos_g::String, imp_g::String, mat_g::String)->1.0,
@@ -77,8 +77,6 @@ const DEFAULT_POPULATION_TYPE = PopulationType(
     # 1.0,
     # 1.0,
     0.0,
-    0,
-    "",
     0.0,
     true,
     0.0,
@@ -96,7 +94,7 @@ const DEFAULT_POPULATION_TYPE = PopulationType(
     ],
     pathogenFractionsWinnerTakesAll,
     weightedResponseWinnerTakesAll,
-    infectionProbabilityWinnerTakesAll,
+    transmissionEfficiencyWinnerTakesAll,
     deNovoResponse,
     Dict{String,ResponseType}([(DEFAULT_RESPONSE_TYPE.id => DEFAULT_RESPONSE_TYPE)])
 )